+++ src/2.0/rfc7591/mod.rs
@@ -1,0 +1,5 @@
+//! Module dedicated to OAuth 2.0 Dynamic Client Registration.
+//!
+//! Refs: <https://datatracker.ietf.org/doc/html/rfc7591>
+
+pub mod client_registration;