+++ src/2.0/rfc8414/mod.rs
@@ -1,0 +1,5 @@
+//! Module dedicated to OAuth 2.0 Authorization Server Metadata.
+//!
+//! Refs: <https://datatracker.ietf.org/doc/html/rfc8414>
+
+pub mod server_metadata;