+++ src/2.0/rfc9728/mod.rs
@@ -1,0 +1,5 @@
+//! Module dedicated to OAuth 2.0 Protected Resource Metadata.
+//!
+//! Refs: <https://datatracker.ietf.org/doc/html/rfc9728>
+
+pub mod resource_metadata;